Moved the server side contact code out of the UserAPIController and into its own ContactAPIController file. #1523
